+++ backend/src/models/pageAttribute.model.ts
@@ -6,13 +6,13 @@ import { modelBuild } from '../utils/model.utils'
 export const PAGE_ATTRIBUTE_CONSTRAINTS = {
   id: {
     nicename: 'ID',
-    type: DataTypes.BIGINT,
+    type: DataTypes.INTEGER,
     autoIncrement: true,
     primaryKey: true
   },
   page_id: {
     nicename: 'Page',
-    type: DataTypes.BIGINT,
+    type: DataTypes.INTEGER,
     required: true
   },
   key: {
